Move logic to read GCS build results to gcs package
