+++ testgrid/cmd/updater/main_test.go
@@ -26,85 +26,6 @@ import (
 	"k8s.io/test-infra/testgrid/state"
 )
 
-func TestParseSuitesMeta(t *testing.T) {
-	cases := []struct {
-		name      string
-		input     string
-		context   string
-		timestamp string
-		thread    string
-		empty     bool
-	}{
-
-		{
-			name:  "not junit",
-			input: "./started.json",
-			empty: true,
-		},
-		{
-			name:  "forgot suffix",
-			input: "./junit",
-			empty: true,
-		},
-		{
-			name:  "basic",
-			input: "./junit.xml",
-		},
-		{
-			name:    "context",
-			input:   "./junit_hello world isn't-this exciting!.xml",
-			context: "hello world isn't-this exciting!",
-		},
-		{
-			name:    "numeric context",
-			input:   "./junit_12345.xml",
-			context: "12345",
-		},
-		{
-			name:    "context and thread",
-			input:   "./junit_context_12345.xml",
-			context: "context",
-			thread:  "12345",
-		},
-		{
-			name:      "context and timestamp",
-			input:     "./junit_context_20180102-1234.xml",
-			context:   "context",
-			timestamp: "20180102-1234",
-		},
-		{
-			name:      "context thread timestamp",
-			input:     "./junit_context_20180102-1234_5555.xml",
-			context:   "context",
-			timestamp: "20180102-1234",
-			thread:    "5555",
-		},
-	}
-
-	for _, tc := range cases {
-		actual := parseSuitesMeta(tc.input)
-		switch {
-		case actual == nil && !tc.empty:
-			t.Errorf("%s: unexpected nil map", tc.name)
-		case actual != nil && tc.empty:
-			t.Errorf("%s: should not have returned a map: %v", tc.name, actual)
-		case actual != nil:
-			for k, expected := range map[string]string{
-				"Context":   tc.context,
-				"Thread":    tc.thread,
-				"Timestamp": tc.timestamp,
-			} {
-				if a, ok := actual[k]; !ok {
-					t.Errorf("%s: missing key %s", tc.name, k)
-				} else if a != expected {
-					t.Errorf("%s: %s actual %s != expected %s", tc.name, k, a, expected)
-				}
-			}
-		}
-	}
-
-}
-
 func TestExtractRows(t *testing.T) {
 	cases := []struct {
 		name     string
