+++ testgrid/util/gcs/BUILD.bazel
@@ -2,7 +2,10 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
 
 go_library(
     name = "go_default_library",
-    srcs = ["gcs.go"],
+    srcs = [
+        "gcs.go",
+        "read.go",
+    ],
     importpath = "k8s.io/test-infra/testgrid/util/gcs",
     visibility = [
         "//prow/gcsupload:__subpackages__",
@@ -10,14 +13,20 @@ go_library(
         "//testgrid:__subpackages__",
     ],
     deps = [
+        "//testgrid/metadata:go_default_library",
+        "//testgrid/metadata/junit:go_default_library",
         "//vendor/cloud.google.com/go/storage:go_default_library",
+        "//vendor/google.golang.org/api/iterator:go_default_library",
         "//vendor/google.golang.org/api/option:go_default_library",
     ],
 )
 
 go_test(
     name = "go_default_test",
-    srcs = ["gcs_test.go"],
+    srcs = [
+        "gcs_test.go",
+        "read_test.go",
+    ],
     embed = [":go_default_library"],
 )
