+++ testgrid/util/gcs/read_test.go
@@ -1,0 +1,100 @@
+/*
+Copyright 2018 The Kubernetes Authors.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+package gcs
+
+import (
+	"testing"
+)
+
+func TestParseSuitesMeta(t *testing.T) {
+	cases := []struct {
+		name      string
+		input     string
+		context   string
+		timestamp string
+		thread    string
+		empty     bool
+	}{
+
+		{
+			name:  "not junit",
+			input: "./started.json",
+			empty: true,
+		},
+		{
+			name:  "forgot suffix",
+			input: "./junit",
+			empty: true,
+		},
+		{
+			name:  "basic",
+			input: "./junit.xml",
+		},
+		{
+			name:    "context",
+			input:   "./junit_hello world isn't-this exciting!.xml",
+			context: "hello world isn't-this exciting!",
+		},
+		{
+			name:    "numeric context",
+			input:   "./junit_12345.xml",
+			context: "12345",
+		},
+		{
+			name:    "context and thread",
+			input:   "./junit_context_12345.xml",
+			context: "context",
+			thread:  "12345",
+		},
+		{
+			name:      "context and timestamp",
+			input:     "./junit_context_20180102-1234.xml",
+			context:   "context",
+			timestamp: "20180102-1234",
+		},
+		{
+			name:      "context thread timestamp",
+			input:     "./junit_context_20180102-1234_5555.xml",
+			context:   "context",
+			timestamp: "20180102-1234",
+			thread:    "5555",
+		},
+	}
+
+	for _, tc := range cases {
+		actual := parseSuitesMeta(tc.input)
+		switch {
+		case actual == nil && !tc.empty:
+			t.Errorf("%s: unexpected nil map", tc.name)
+		case actual != nil && tc.empty:
+			t.Errorf("%s: should not have returned a map: %v", tc.name, actual)
+		case actual != nil:
+			for k, expected := range map[string]string{
+				"Context":   tc.context,
+				"Thread":    tc.thread,
+				"Timestamp": tc.timestamp,
+			} {
+				if a, ok := actual[k]; !ok {
+					t.Errorf("%s: missing key %s", tc.name, k)
+				} else if a != expected {
+					t.Errorf("%s: %s actual %s != expected %s", tc.name, k, a, expected)
+				}
+			}
+		}
+	}
+
+}
